refactor(parsers): canonicalize mixed-think tool-handoff naming with step_35 compatibility
